Fixes a bug

+++ aligners/bwt/bwt.c
@@ -3248,7 +3248,7 @@ void insert_seeds_and_merge(array_list_t *mapping_list, linked_list_t ***cals_li
     int strand = region->strand;    
     //printf("Insert Region (%i)[%i:%lu|%i-%i|%lu]\n", region->strand, region->chromosome_id, 
     // region->start, region->seq_start, region->seq_end, region->end);
-    my_cp_list_append_linked_list(cals_list[strand][chromosome_id], region, (size_t) max_cal_distance, 1000);
+    my_cp_list_append_linked_list(cals_list[strand][chromosome_id], region, max_cal_distance, 1000);
 
     region_bwt_free(region);
   }
